Remove references to stack and heap
I was under the impression that arrays are always stored on the heap, however a quick google search reveals this not to be the case. In short, if the array was declared within a function, then it will be stored on the stack.

However, whether the variables are on the stack or on the heap has little to do with what we are trying to teach here, so we will simply remove references to those and the overall point remains the same.