Add Stack and Heap data types
After some more research, it does appear to be the case that arrays are in fact always stored on the heap. What may be stored on the stack is the reference to the array.

With this commit we store our list of declared variables on the stack, and the values of the reference types on a custom heap data type.

Final output is bugged and needs fixing.