Add AltListLockLock1<T1, T2> data structure
This represents a list that alternates between types T1 and T2. The first "Lock" denotes that the first element is "locked", i.e. must be of type T1. The "Lock1" denotes that the last element is also "locked" and specifically to type T1.

This data structure is designed in such a way that it is impossible to break those rules. This will make it much easier to control evaluables that need to be separated by operators.