Add string constructor
This allows us to construct a string from a char array. This is important because up until now the only way to create a new string was with a literal, and since strings are immutable, there was no way to create a string whose value wasn't known at compile time.

On the other hand, it is easy to create and mutate an array of characters at runtime. Therefore, by adding the ability to create a string from a character array, we also add the ability to create a string whose value only becomes known at runtime. When we do this, a copy of all the character array elements is made, and inserted into the string (thus both the character array and the string will be on the heap, as separate entities).