Replace LineNumber with LineNumberInfo struct
Instead of merely having the line number for each line, we now have three three distinct line numbers:
1) the line number at which the LINE starts
2) the line number of the first non-whitespace character
3) the line number of the last non-whitespace character

Note that we are using two different definitions of "line" here. Line (in lowercase) is defined as a string of text separated by newline characters. LINE (in capitals) is defined as a statement to execute, as per the Line class, separated by either semicolons or curly braces as appropriate.