Perform static initialisation analysis for nested statements
if, if-else, while, and for statements may initialise variables that are declared outside of their scope. However, the compiler should uninitialise them after, as there is no guarantee that that code will ever run at runtime. The only exception is in the case of the if-else statement: in that case, if a variable is initialised in both the if and the else parts of the statement, then we leave it initialised.