Add C# -> Java test converter
This is a C# program that finds each .cs test file, makes a copy of it, renames it to a .java file, then replaces all "bool" with "boolean", all "string" with "String", and all "Length" with "length". Finally, it prints out which files contain "length", so that we can manually add brackets "()" for string lengths.