Fix stack scrollbar overflow bug
This was a bug that was only present in Chrome-based browsers. In Firefox it was already working fine. The solution makes the right-padding for very long stack variable names not the best looking, but it's the best that can be done without getting extremely hacky. On Firefox it looks great though.